Check for an existing signed-in user
In your activity's onStart method, check if a user has already signed in to your app with Google.

f GoogleSignIn.getLastSignedInAccount returns a GoogleSignInAccount object (rather than null),
the user has already signed in to your app with Google.
Update your UI accordingly—that is, hide the sign-in button, launch your main activity,
or whatever is appropriate for your app.

If GoogleSignIn.getLastSignedInAccount returns null, the user has not yet signed in to your app with Google.
Update your UI to display the Google Sign-in button (or other fail message).