Add logout functionality.
Edited mainActivity with logout button and code
(added listener for logout button, signout method and updateUi changes)
The logged activity is called by a new button: show more info,
because the logout method needs to be in mainActivity.

Edited logged activity with a back button to return in main activity
(for logout-login).
The intent extra for switch activities is collapsed in an
Arraylist<String> instead of many extra-strings.

For more info about google signout method:
https://developers.google.com/identity/sign-in/android/disconnect